move all the EFI variable stuff to a module
this makes the tests kind of legible..
